Move `Strings` to `io.spine.base.string`
